pyproject.toml
[project]
name = "3b1b-at-home"
version = "0.1.0"
description = "Automated math education video pipeline"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "filelock>=3.13",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.29",
    "manim>=0.18.0",
    "kokoro>=0.9.2",
    "soundfile>=0.12",
    "numpy>=1.26",
    "google-generativeai>=0.8",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
